third_party/blink/web_tests/platform/generic/fast/forms/setCustomValidity-existence-expected.txt

Existence tests for .setCustomValidity property

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Existence of .setCustomValidity
PASS typeof document.getElementsByTagName("input")[0].setCustomValidity is "function"
PASS typeof document.getElementsByTagName("button")[0].setCustomValidity is "function"
PASS typeof document.getElementsByTagName("fieldset")[0].setCustomValidity is "function"
PASS typeof document.getElementsByTagName("select")[0].setCustomValidity is "function"
PASS typeof document.getElementsByTagName("textarea")[0].setCustomValidity is "function"
PASS typeof document.getElementsByTagName("output")[0].setCustomValidity is "function"
PASS typeof document.getElementsByTagName("object")[0].setCustomValidity is "function"
PASS successfullyParsed is true

TEST COMPLETE

Test
